+++ components/service/block_storage/block_store/client/block_storage_client.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022, Arm Limited. All rights reserved.
+ * Copyright (c) 2022-2023, Arm Limited. All rights reserved.
  *
  * SPDX-License-Identifier: BSD-3-Clause
  *
@@ -186,7 +186,7 @@ static psa_status_t block_storage_client_close(void *context,
 static psa_status_t block_storage_client_read(void *context,
 	uint32_t client_id,
 	storage_partition_handle_t handle,
-	uint32_t lba,
+	uint64_t lba,
 	size_t offset,
 	size_t buffer_size,
 	uint8_t *buffer,
@@ -252,7 +252,7 @@ static psa_status_t block_storage_client_read(void *context,
 static psa_status_t block_storage_client_write(void *context,
 	uint32_t client_id,
 	storage_partition_handle_t handle,
-	uint32_t lba,
+	uint64_t lba,
 	size_t offset,
 	const uint8_t *data,
 	size_t data_len,
@@ -321,7 +321,7 @@ static psa_status_t block_storage_client_write(void *context,
 static psa_status_t block_storage_client_erase(void *context,
 	uint32_t client_id,
 	storage_partition_handle_t handle,
-	uint32_t begin_lba,
+	uint64_t begin_lba,
 	size_t num_blocks)
 {
 	struct block_storage_client *this_context = (struct block_storage_client *)context;
@@ -334,7 +334,7 @@ static psa_status_t block_storage_client_erase(void *context,
 
 	req_msg.handle = handle;
 	req_msg.begin_lba = begin_lba;
-	req_msg.num_blocks = (uint32_t)num_blocks;
+	req_msg.num_blocks = num_blocks;
 
 	rpc_call_handle call_handle =
 		rpc_caller_begin(this_context->client.caller, &req_buf, req_len);
