+++ components/service/block_storage/provider/serializer/block_storage_serializer.h
@@ -42,14 +42,14 @@ struct block_storage_serializer {
 	/* Operation: read */
 	rpc_status_t (*deserialize_read_req)(const struct call_param_buf *req_buf,
 		storage_partition_handle_t *handle,
-		uint32_t *lba,
+		uint64_t *lba,
 		size_t *offset,
 		size_t *len);
 
 	/* Operation: write */
 	rpc_status_t (*deserialize_write_req)(const struct call_param_buf *req_buf,
 		storage_partition_handle_t *handle,
-		uint32_t *lba,
+		uint64_t *lba,
 		size_t *offset,
 		const uint8_t **data,
 		size_t *data_len);
@@ -60,7 +60,7 @@ struct block_storage_serializer {
 	/* Operation: erase */
 	rpc_status_t (*deserialize_erase_req)(const struct call_param_buf *req_buf,
 		storage_partition_handle_t *handle,
-		uint32_t *begin_lba,
+		uint64_t *begin_lba,
 		size_t *num_blocks);
 };
